magit-refresh-status: better status lines
When there is a push remote use "Push Remote" resp. "Pull Remote" as
the keywords.  These keywords are longer than the previously longest
keyword.  I don't want to add more spaces than needed when aligning so
the alignment column depends on whether there is a push remote or not.
For that reason add `magit-insert-status-line', the code would become
quite unreadable without this.